The sample and the library has the same package, so sometimes the IDE could not recognize the id of the views
