+++ kappuccino/build.gradle
@@ -27,8 +27,7 @@ android {
         abortOnError false
     }
 
-    sourceSets { main { java.srcDirs = ['src/main/kotlin'] } }
-    buildToolsVersion '27.0.3'
+    sourceSets { main { java.srcDirs += 'src/main/kotlin' } }
 }
 
 dependencies {
