+++ kappuccino/src/main/kotlin/br/com/concretesolutions/kappuccino/custom/compoundDrawable/ClickDrawableAction.kt
@@ -26,10 +26,10 @@ class ClickDrawableAction(@Location private val drawableLocation: Int) : ViewAct
         @Retention(AnnotationRetention.SOURCE)
         annotation class Location
 
-        const val Left = 0L
-        const val Top = 1L
-        const val Right = 2L
-        const val Bottom = 3L
+        const val Left = 0
+        const val Top = 1
+        const val Right = 2
+        const val Bottom = 3
     }
 
     override fun getDescription(): String {
