+++ sample/build.gradle
@@ -25,9 +25,8 @@ android {
     }
 
     sourceSets {
-        androidTest.java.srcDirs = ['src/androidTest/kotlin']
+        androidTest.java.srcDirs += 'src/androidTest/kotlin'
     }
-    buildToolsVersion '27.0.3'
 }
 
 dependencies {
