+++ sample/src/androidTest/kotlin/br/com/concretesolutions/kappuccino/test/ManipulateTextActivityTest.kt
@@ -2,13 +2,13 @@ package br.com.concretesolutions.kappuccino.test
 
 import android.support.test.rule.ActivityTestRule
 import android.support.test.runner.AndroidJUnit4
-import br.com.concretesolutions.kappuccino.ManipulateTextActivity
-import br.com.concretesolutions.kappuccino.R
+import br.com.concretesolutions.kappuccino.sample.R
 import br.com.concretesolutions.kappuccino.actions.ClickActions.click
 import br.com.concretesolutions.kappuccino.assertions.VisibilityAssertions.displayed
 import br.com.concretesolutions.kappuccino.matchers.drawable.DrawablePosition.End
 import br.com.concretesolutions.kappuccino.matchers.drawable.DrawablePosition.Start
 import br.com.concretesolutions.kappuccino.matchers.drawable.DrawablePosition.Top
+import br.com.concretesolutions.kappuccino.sample.ManipulateTextActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
