+++ sample/src/androidTest/kotlin/br/com/concretesolutions/kappuccino/test/VisibilityActivityTest.kt
@@ -3,13 +3,14 @@ package br.com.concretesolutions.kappuccino.test
 import android.support.test.espresso.intent.Intents
 import android.support.test.rule.ActivityTestRule
 import android.support.test.runner.AndroidJUnit4
-import br.com.concretesolutions.kappuccino.MainActivity
-import br.com.concretesolutions.kappuccino.R
-import br.com.concretesolutions.kappuccino.VisibilityAssertionsActivity
 import br.com.concretesolutions.kappuccino.actions.ClickActions.click
 import br.com.concretesolutions.kappuccino.assertions.VisibilityAssertions.displayed
 import br.com.concretesolutions.kappuccino.assertions.VisibilityAssertions.notDisplayed
-import br.com.concretesolutions.kappuccino.custom.intent.IntentMatcherInteractions.matchIntent
+import br.com.concretesolutions.kappuccino.custom.intent.IntentMatcherInteractions.sentIntent
+import br.com.concretesolutions.kappuccino.custom.intent.IntentMatcherInteractions.stubIntent
+import br.com.concretesolutions.kappuccino.sample.MainActivity
+import br.com.concretesolutions.kappuccino.sample.R
+import br.com.concretesolutions.kappuccino.sample.VisibilityAssertionsActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -17,7 +18,8 @@ import org.junit.runner.RunWith
 @RunWith(AndroidJUnit4::class)
 class VisibilityActivityTest {
 
-    @Rule @JvmField
+    @Rule
+    @JvmField
     var mActivityRule = ActivityTestRule<VisibilityAssertionsActivity>(VisibilityAssertionsActivity::class.java, false, true)
 
     @Test
@@ -109,9 +111,9 @@ class VisibilityActivityTest {
     fun match_intentTo_mainActivity() {
         Intents.init()
 
-        matchIntent {
+        stubIntent {
             className(MainActivity::class.java.name)
-            result {
+            respondWith {
                 ok()
             }
         }
@@ -120,7 +122,7 @@ class VisibilityActivityTest {
             id(R.id.btn_to_main)
         }
 
-        matchIntent {
+        sentIntent {
             className(MainActivity::class.java.name)
         }
