--- sample/src/main/java/br/com/concretesolutions/kappuccino/RecyclerViewAdapter.java
@@ -1,58 +1,0 @@
-package br.com.concretesolutions.kappuccino;
-
-import android.content.Context;
-import android.support.v7.widget.RecyclerView;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapter.ViewHolder>{
-
-    String[] SubjectValues;
-    Context context;
-    View view1;
-    ViewHolder viewHolder1;
-
-    public RecyclerViewAdapter(Context context1,String[] SubjectValues1){
-
-        SubjectValues = SubjectValues1;
-        context = context1;
-    }
-
-    public static class ViewHolder extends RecyclerView.ViewHolder{
-
-        public TextView textView;
-
-        public ViewHolder(View v){
-
-            super(v);
-
-            textView = (TextView)v.findViewById(R.id.subject_textview);
-        }
-    }
-
-    @Override
-    public RecyclerViewAdapter.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType){
-
-        view1 = LayoutInflater.from(context).inflate(R.layout.recyclerview_items,parent,false);
-
-        viewHolder1 = new ViewHolder(view1);
-
-        return viewHolder1;
-    }
-
-    @Override
-    public void onBindViewHolder(ViewHolder holder, int position){
-
-        holder.textView.setText(SubjectValues[position]);
-//        if (position == 0)
-//            holder.textView.setVisibility(View.INVISIBLE);
-    }
-
-    @Override
-    public int getItemCount(){
-
-        return SubjectValues.length;
-    }
-}
