+++ sample/src/main/java/br/com/concretesolutions/kappuccino/sample/ManipulateTextActivity.kt
@@ -1,7 +1,7 @@
-package br.com.concretesolutions.kappuccino
-import android.support.v7.app.AppCompatActivity
+package br.com.concretesolutions.kappuccino.sample
 import android.os.Bundle
 import android.support.v4.content.ContextCompat
+import android.support.v7.app.AppCompatActivity
 import android.view.View
 import android.widget.Button
 import android.widget.TextView
